Move tenant-related utils to common folder (#1184)
(cherry picked from commit f815e3c)
